Make EmptyQueryResultsMessage into a component
